Tools: use much more cleaner way of calling tools
Eliminates extra classes

Add UI interstitial name for the Tool
