+++ requirements.txt
@@ -1,10 +1,11 @@
 aiofiles==24.1.0
-aiosqlite==0.20.0
 google-generativeai >= 0.7.2
 google_labs_html_chunker==0.0.5
 gradio_client==1.1.0
 jsonpickle==3.2.2
+motor==3.5.1
 py-cord >= 2.6.0
+pymongo==4.8.0
 python-dotenv==1.0.0
 PyYAML==6.0.1
 # wavelink==3.4.1
